Update script and add linting step
We need to wait after searching each organisation for files,
to avoid hitting the search API rate limits.

We're making an assumption that files containing the three required fields for OpenAPI
objects, `openapi`, `info` and `servers`,[1] and that are `yaml` or `json` files, are likely
to be OpenAPI documents.

We decided to ignore documents in archived repositories and in test
directories as these are unlikely to be in use or describing real APIs.

To get the raw content to lint, we need to change the urls of the documents
we find to `raw.githubusercontent.com`.

[1] https://swagger.io/specification/